+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Updated `AddressContainer` to validate address on update if `shouldHandleAddressChangeOnMount` is passed.
+
 ## [4.5.0] - 2020-02-14
 
 ### Changed